+++ src/crud_views/lib/formsets/render_tree.py
@@ -93,6 +93,7 @@ class XFormSet(BaseModel, arbitrary_types_allowed=True):
     level: int
     prefix_key: str
     formset: Any
+    cv_active: bool = True
     instance: BaseFormSet
     management_form: Any
     forms: List[XForm] = Field(default_factory=lambda: list())
@@ -189,6 +190,11 @@ class XFormSet(BaseModel, arbitrary_types_allowed=True):
     @property
     def json_data(self) -> str:
         return json.dumps(self.data)
+
+    @property
+    def cv_toggle_field(self) -> str:
+        conditional = self.formset.conditional
+        return conditional.toggle.field_name() if conditional is not None else ""
 
     def is_valid(self) -> Iterable[Tuple[Any, bool]]:
         yield self, self.instance.is_valid()